changed user media files structure to a folder for each item and list & refactored all backend media-related functions & API endpoints according to it.
new structure:
└── userID/
	├── list1ID > its cover& logos
	│	  ├── item1ID > its cover, poster, images & logos
	│	  └── item2ID ...
	└── list2ID ...
with a 'template' prefix for list's field's template logos only

old structure:
└── userID/
	├──lists > all lists' covers
	├──items > all items' covers, posters and images
	└──logos > logos of items and lists

- in the old structure (was mostly just for dev):
	- deleting logos or when editing logos fields and deleting a logo from them will have god-knows how much  (at least O(n^2)) time complexity:
		- it needs to iterate over all to-be-deleted items to check collect logo_paths -> O(n) where n is the number of items
		- it needs to iterate over all to-be-deleted logos to delete them -> O(nm) ,and m is the number of logos which can dramatically exceed the number of items
		- then inside that iteration it needs to iterate over all the logos paths of all items to check that the logo is not used by any item -> O(nmn) = O(n^2 m)
		- iterate over all templates' logos (usually a short list) to check that it isn't one of them - > still O(nmn) or O(n^2 m)
	- deleting an item:
		- should iterate over all logo_paths, poster, cover & images of the item to delete them. and do the same check -> O(nm)
	- deleting a list:
		- should iterate over all logo_paths & cover of the list to delete them  -> O(m)
		- iterate over all all items and collect logo_paths then delete them -> O(nmm) = O(n m^2)
- in the new structure:
	- when editing logos and deleting one of them, inside:
		- item:
			-  it needs to iterate all to-be-deleted logos to delete them - O(m) - where m is is the logos,
			- in the **same** iteration if the logo has a 'template' prefix it will skip it because it is a field's template - O(m)
		- list:
			-  it needs to iterate all to-be-deleted logos - O(m)
			-  the one will have a 2nd iteration to check if the logo is used by an item, if so ignore it. - O(nm)
	- when deleting:
		- lists: just delete the list's folder (which includes items' data) -> O(n)  where n is the number of lists
		- items: just delete the item's folder -> O(n) where n is the number of items to delete

This will have a better performance, and makes logos' deleting simpler

+++ server/src/routes/images.ts
@@ -4,10 +4,9 @@ import { items_images } from '@prisma/client';
 import express from 'express';
 import formidable from 'formidable';
 import { validate as uuidValidate } from 'uuid';
+import deleteImage from '../utils/handlers/deleteImageFn';
 import handleFileSaving from '../utils/handlers/handleFileSaving';
-import userMediaFoldersPath from '../utils/userMediaFoldersPath';
-import path from 'path';
-import fs from 'fs';
+import userMediaRoot from '../utils/userMediaRoot';
 
 const imagesRouter = express.Router();
 
@@ -29,10 +28,10 @@ imagesRouter.post('/:item_id', async (req, res) => {
 
         if (!files?.image_path?.[0]) return res.status(400).json({ message: 'Bad Request: No Image Provided' })
 
-        const userMediaRoot = userMediaFoldersPath(user_id);
+        const ItemMediaRoot = userMediaRoot(user_id, item.list_id, item.id);
         let data = {} as items_images;
 
-        data.image_path = await handleFileSaving(files.image_path[0], userMediaRoot.items);
+        data.image_path = await handleFileSaving(files.image_path[0], ItemMediaRoot);
         data.user_id = user_id
         data.item_id = item.id
         data.title = fields?.title?.[0] ?? null
@@ -81,10 +80,11 @@ imagesRouter.delete('/:id', async (req, res) => {
         const image = await prisma.items_images.findUnique({ where: { id, user_id } })
         if (!image) return res.status(404).json({ message: `Bad Image ID, Image doesn't exist` })
 
-        const userMediaRoot = userMediaFoldersPath(user_id);
-        
-        const filePath = path.join(userMediaRoot.items, image.image_path)
-        await fs.promises.unlink(filePath)
+        const item = await prisma.items.findUnique({ where: { id: image.item_id, user_id } })
+        if (!item) return res.status(404).json({ message: `Bad Item ID, Item doesn't exist` })
+
+        const ItemMediaRoot = userMediaRoot(user_id, item.list_id, item.id);
+        await deleteImage(ItemMediaRoot, image.image_path)
 
         await prisma.items_images.delete({
             where: { id, user_id }
